requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.*
factory-boy==3.3.1
django-filter==24.3
drf-yasg==1.21.8
XlsxWriter==3.2.0
celery==5.4.0
redis==5.2.0
# optional
#flower==2.0.1
#django-celery==3.3.1
#django-celery-results==2.5.1
